# backend-api/docs/openapiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chemistry Game Suite API",
    "description": "API cho bộ sưu tập trò chơi hóa học dựa trên OpenAPI 3.0, hỗ trợ quản lý phản ứng, ô chữ, nguyên tố, và phiên chơi. \nTìm hiểu thêm về Swagger tại [https://swagger.io](https://swagger.io).",
    "termsOfService": "https://yourchemistrygame.com/terms/",
    "contact": {
      "email": "giaphuc2004hg@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.12"
  },
...
  "servers": [
    {
      "url": "https://ct313hm01-project-gphuc04.onrender.com/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Lỗi dữ liệu đầu vào"
          },
          "error": {
            "type": "string",
            "example": "Invalid input"
          }
        },
        "required": ["code", "message"]
      },
...
    }
  },
  "paths": {
...
    "/crossword-puzzles": {
      "get": {
        "tags": ["Crossword Puzzles"],
        "summary": "Lấy danh sách tất cả các câu đố ô chữ",
        "description": "API để lấy danh sách tất cả các câu đố ô chữ, yêu cầu xác thực JWT.",
        "operationId": "getAllCrosswordPuzzles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách câu đố ô chữ thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách câu đố ô chữ thành công"
                    },
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "puzzle_id": {
                            "type": "string",
                            "example": "puz123"
                          },
                          "title": {
                            "type": "string",
                            "example": "Chemistry Basics"
                          },
                          "grid_size": {
                            "type": "integer",
                            "example": 10
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Crossword Puzzles"],
        "summary": "Tạo mới một câu đố ô chữ",
        "description": "API để tạo một câu đố ô chữ mới, yêu cầu xác thực JWT.",
        "operationId": "createCrosswordPuzzle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title_vi": {
                    "type": "string",
                    "example": "Cơ bản Hóa học",
                    "description": "Tiêu đề tiếng Việt của câu đố"
                  },
                  "title_en": {
                    "type": "string",
                    "example": "Chemistry Basics",
                    "description": "Tiêu đề tiếng Anh của câu đố"
                  },
                  "grid_size": {
                    "type": "integer",
                    "example": 15,
                    "description": "Kích thước lưới ô chữ (ví dụ: 15 = 15x15)"
                  },
                  "grid_data": {
                    "type": "object",
                    "example": {
                      "cells": [
                        ["C", "H", "O"],
                        ["", "H", ""],
                        ["", "", "N"]
                      ]
                    },
                    "description": "Dữ liệu ô chữ (gồm các chữ cái hoặc ô trống)"
                  },
                  "clues_across": {
                    "type": "array",
                    "description": "Danh sách gợi ý hàng ngang",
                    "items": {
                      "type": "object",
                      "properties": {
                        "clue": {
                          "type": "string",
                          "example": "Nguyên tố có ký hiệu là H"
                        },
                        "answer": {
                          "type": "string",
                          "example": "Hydrogen"
                        },
                        "position": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  },
                  "clues_down": {
                    "type": "array",
                    "description": "Danh sách gợi ý hàng dọc",
                    "items": {
                      "type": "object",
                      "properties": {
                        "clue": {
                          "type": "string",
                          "example": "Nguyên tố có ký hiệu là O"
                        },
                        "answer": {
                          "type": "string",
                          "example": "Oxygen"
                        },
                        "position": {
                          "type": "integer",
                          "example": 2
                        }
                      }
                    }
                  },
                  "solution": {
                    "type": "object",
                    "example": {
                      "cells": [
                        ["C", "H", "O"],
                        ["", "H", ""],
                        ["", "", "N"]
                      ]
                    },
                    "description": "Đáp án hoàn chỉnh của ô chữ"
                  },
                  "difficulty_level": {
                    "type": "string",
                    "enum": ["easy", "medium", "hard", "expert"],
                    "example": "medium",
                    "description": "Độ khó của câu đố"
                  },
                  "is_public": {
                    "type": "boolean",
                    "example": true,
                    "description": "Câu đố có công khai hay không"
                  }
                },
                "required": [
                  "title_vi",
                  "title_en",
                  "grid_size",
                  "grid_data",
                  "clues_across",
                  "clues_down",
                  "solution"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo câu đố ô chữ thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "POST, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo câu đố ô chữ thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "puzzle_id": {
                          "type": "integer",
                          "example": 42
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi dữ liệu đầu vào không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/crossword-puzzles/{Id}": {
      "get": {
        "tags": ["Crossword Puzzles"],
        "summary": "Lấy thông tin câu đố ô chữ theo ID",
        "description": "API để lấy thông tin chi tiết của một câu đố ô chữ dựa trên puzzleId, yêu cầu xác thực JWT.",
        "operationId": "getCrosswordPuzzleById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "ID của câu đố ô chữ cần lấy thông tin",
            "schema": {
              "type": "string",
              "example": "puz123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin câu đố ô chữ thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy thông tin câu đố ô chữ thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "puzzle_id": {
                          "type": "string",
                          "example": "puz123"
                        },
                        "title": {
                          "type": "string",
                          "example": "Chemistry Basics"
                        },
                        "grid_size": {
                          "type": "integer",
                          "example": 10
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID câu đố ô chữ không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy câu đố ô chữ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}